Fixed comments in PR #1

+++ booking-service/src/main/kotlin/repository/mongo/MongoBookingRepository.kt
@@ -22,10 +22,16 @@ class MongoBookingRepository(database: MongoDatabase) : BookingRepositoryInterfa
     }
 
     override suspend fun findById(id: String): Booking? = withContext(Dispatchers.IO) {
+        if(!ObjectId.isValid(id)) {
+            throw IllegalArgumentException("Invalid booking ID format: $id")
+        }
         collection.find(Filters.eq("_id", ObjectId(id))).first()?.toBooking()
     }
 
     override suspend fun deleteById(id: String): Booking? = withContext(Dispatchers.IO) {
+        if(!ObjectId.isValid(id)) {
+            throw IllegalArgumentException("Invalid booking ID format: $id")
+        }
         collection.findOneAndDelete(Filters.eq("_id", ObjectId(id)))?.toBooking()
     }
